Removed all FRW references.
The only substantive use FRW was used in are the HistWrappers. This behavioris going to need to be readded in MAT-MINERvA if we want it.

I believe there was nothing stopping FluxSystematics from going to
MAT-MINERvA yesterday.

Also modified the dict's and build's CMakeLists. Don't know what
I was doing with that, but it worked.